backend/requirements.txt
# FastAPI and ASGI Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Pydantic for data validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Database - SQLAlchemy and SQLite
sqlalchemy==2.0.25
aiosqlite==0.19.0
alembic==1.13.0

# Database - Neo4j
neo4j==5.15.0

# Database - Redis
redis==4.6.0
hiredis==2.3.2

# Celery for background tasks
celery[redis]==5.3.4

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
cryptography==41.0.7

# LLM Providers
ollama==0.1.6
anthropic==0.18.0
openai==1.10.0

# LangChain and RAG
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.5
sentence-transformers==2.3.1
faiss-cpu==1.8.0

# Document Processing
pymupdf==1.23.8
python-docx==1.1.0
unstructured==0.11.0

# Audio Processing and Speech-to-Text
openai-whisper==20231117
pyannote.audio==3.1.1
soundfile==0.12.1
librosa==0.10.1

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
uuid6==2023.5.2

# Logging and Monitoring
structlog==24.1.0
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
faker==22.0.0

# Development
black==23.12.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
